ai-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
groq==0.4.2
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0